// src/app/models/Persona2.ts
export class Persona2 {
  id_persona: number = 0;
  cedula: string = "";
  primer_nombre: string = "";
  segundo_nombre: string = "";
  primer_apellido: string = "";
  segundo_apellido: string = "";
  correo: string = "";
  direccion: string = "";
  celular: string = "";
  visible: boolean = true;
}